[front] refactor: move ExternalLink in its own file
